remove anything that might conflict with other ppls changes
